old_src/requirements/requirements.txt
# API Framework
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.7
pydantic>=2.6.0

# Database
SQLAlchemy>=2.0.27
psycopg2-binary>=2.9.9
alembic>=1.13.1

# HTTP Client
httpx>=0.27.0

# Environment and Configuration
python-dotenv>=1.0.1
python-json-logger>=2.0.7

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
tenacity>=8.2.3
typing_extensions>=4.9.0
